prettyping.sh: Improving the legend drawing.
If there is a backgroud change, and the next character being printed does not
fit the current line, then the next line will be filled with that background
color. This is not desired and looks ugly, so had to be fixed. Also implemented
some logic to avoid changing the line in the middle of a legend/label.