Added technicolor_gateway_adsl_info.py
This is a script to grab ADSL status information from a TG580 Wireless ADSL2+
Gateway and display such information at the command-line.

This script uses pyparsing to implement a very limited portion of the
JavaScript language, but just enough to parse the device output.

It also includes routines to generate a human-readable version of huge numbers
(using either decimal or binary prefixes) and a human-readable version of the
amount of seconds.