Adding a Gimp plug-in to auto-position layers based on the layer name.
Suppose we have several pictures named `foobar_1_2.png`, where `1` and `2` are
different numbers for each picture, and represent the X and Y position in some
arbitrary coordinate system. Then, the user can import all those pictures as
layers into a single image and run this plug-in to position all those layers at
the correct position.

The actual formula is hardcoded, as well as the code to extract the numbers
from the name. The script must be manually edited for each case.